==> scripts/0-packages.R <==
# all packages needed for scripts
library(tidyverse)
library(maps)
library(bibliometrix)
library(ggthemes)
library(ggpubr)
library(ggrepel)
library(plotly)
library(rphylopic)
library(png)
library(cowplot)
library(cartography)
library(sf)
library(data.table)
library(janitor)
